Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -7,7 +7,7 @@ int main(){
     int sum = x+y;
     int minus = x-y;
     int product = x*y;
-    int divide = x/y;
+    float divide = x/y;
     if(z=='+'){
         printf("%d",sum);
     }
